requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-factory-boy==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.4
factory-boy==3.2.1
Faker==15.3.4
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
